Integration/product-price-category (#55)
* :hammer: integrate price

* :art: create product

* :hammer: update categories on product

* :art: remove duplicates

* :hammer: update listing of products with categories

* :fire: remove duplicate model for product

* :hammer: add dialog box for deleting product

* Update UI/Api/ProductService.cs


* :fire: remove unused return

---------

+++ UI/Pages/Products/Views/MainView.axaml.cs
@@ -10,6 +10,7 @@ using Pos.Models;
 using ReactiveUI;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
+using AvaloniaDialogs.Views;
 
 namespace Pos.Pages.Products;
 
@@ -33,7 +34,18 @@ public partial class MainView : ReactiveUserControl<MainViewModel>
 
     private async void DeleteButton_Click(object sender, RoutedEventArgs e)
     {
-        await ViewModel?.DeleteProductById(Convert.ToInt32(((Button)sender).Tag));
+        TwofoldDialog dialog = new()
+        {
+            Message = "Delete this Product?",
+            PositiveText = "Delete",
+            NegativeText = "Cancel"
+        };
+        dialog.FindControl<Button>("PositiveButton")!.Classes.Add("Danger");
+        if ((await dialog.ShowAsync()).GetValueOrDefault(false))
+        {
+            await ViewModel?.DeleteProductById(Convert.ToInt32(((Button)sender).Tag));
+        }
+
     }
 
     private void EditButton_Click(object sender, RoutedEventArgs e)
